+++ Include/LCE/Simulation/Decision/Outcome.h
@@ -43,8 +43,8 @@
 
 #pragma once
 
-#include "LCE/Simulation/EntityId.h"
-#include "LCE/Simulation/Memory.h"
+#include "LCE/Simulation/Entity/EntityId.h"
+#include "LCE/Simulation/Mind/Memory.h"
 
 namespace LCE::Simulation
 {