+++ Include/LCE/Simulation/Mind/Memory.h
@@ -43,7 +43,7 @@
 
 #pragma once
 
-#include "LCE/Simulation/EntityId.h"
+#include "LCE/Simulation/Entity/EntityId.h"
 
 #include <cstdint>
 #include <vector>